move connection handler to its own header file
